refactor(forms): improve some form hooks
