+++ src/app/(client)/hooks/forms/useLoginForm.ts
@@ -4,8 +4,9 @@
  * Thin wrapper around useFormWithSchema for the “Login” screen.
  * No extra helpers needed besides the raw RHF API.
  */
-import { useFormWithSchema } from '@/app/(client)/hooks/forms';
 import { LoginSchema, loginDefaults } from '@/lib/validation';
+
+import { useFormWithSchema } from './useFormWithSchema';
 
 export function useLoginForm(mode: 'onBlur' | 'onChange' | 'onSubmit' = 'onBlur') {
 	return useFormWithSchema(LoginSchema, loginDefaults, mode);
